% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_backup.R
\name{backupPackages}
\alias{backupPackages}
\title{Backup Installed Packages}
\usage{
backupPackages(file_path = "package_backup.csv")
}
\arguments{
\item{file_path}{Character string for backup file location}
}
\value{
Invisible NULL
}
\description{
Creates a backup of installed packages and their versions
}
