% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_checker.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check Package Dependencies}
\usage{
checkDependencies(pkg_name)
}
\arguments{
\item{pkg_name}{Character string of package name}
}
\value{
List with dependency status
}
\description{
Validates package dependencies and their versions
}
