% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checker.R
\name{checkPackageVersion}
\alias{checkPackageVersion}
\title{Check Package Version}
\usage{
checkPackageVersion(pkg_name)
}
\arguments{
\item{pkg_name}{Character string of package name}
}
\value{
List with installed and available versions
}
\description{
Compares installed and available versions of a package
}
