% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checker.R
\name{checkUpdaterVersion}
\alias{checkUpdaterVersion}
\title{Check Package Update Version}
\usage{
checkUpdaterVersion()
}
\value{
List with local and remote versions and update status
}
\description{
Checks if there's a newer version of the package on GitHub
}
