% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_backup.R
\name{restorePackages}
\alias{restorePackages}
\title{Restore Packages from Backup}
\usage{
restorePackages(file_path = "package_backup.csv")
}
\arguments{
\item{file_path}{Character string for backup file location}
}
\value{
List of restoration results
}
\description{
Restores packages from a backup file
}
