% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{updateRpackages}
\alias{updateRpackages}
\title{Update and Rebuild R Packages}
\usage{
updateRpackages(debug_output = FALSE)
}
\arguments{
\item{debug_output}{Logical. If TRUE, uses regular R console output instead of formatted output}
}
\description{
Updates all installed R packages and rebuilds if necessary
}
