% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_updater.R, R/updateR.R
\name{rebuildPackages}
\alias{rebuildPackages}
\title{Rebuild Packages}
\usage{
rebuildPackages()

rebuildPackages()
}
\value{
Named list of rebuild results
}
\description{
Rebuilds packages that were built with a different R version
}
