% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{updateRpackages}
\alias{updateRpackages}
\title{Update and Rebuild R Packages}
\usage{
updateRpackages()
}
\description{
Updates all installed R packages from source, rebuilds outdated ones, and retries using binaries if necessary.
}
